Update CRUD logic

+++ src/controller/item.controller.ts
@@ -7,9 +7,15 @@ const itemService = new ItemService();
 
 const createItem = async (event: APIGatewayEvent) => {
   const itemData = JSON.parse(event.body ?? "{}");
-  const newItem = await itemService.create(itemData);
 
-  return createResponse(StatusCodes.CREATED, newItem);
+  try {
+    const newItem = await itemService.create(itemData);
+    return createResponse(StatusCodes.CREATED, newItem);
+  } catch (error) {
+    return createResponse(StatusCodes.BAD_REQUEST, {
+      message: ErrorMessages.INVALID_DATA,
+    });
+  }
 };
 
 const getItem = async (event: APIGatewayEvent) => {
@@ -21,15 +27,34 @@ const getItem = async (event: APIGatewayEvent) => {
     });
   }
 
-  const item = await itemService.get(itemId);
+  try {
+    const item = await itemService.get(itemId);
 
-  return createResponse(StatusCodes.OK, item);
+    if (!item) {
+      return createResponse(StatusCodes.NOT_FOUND, {
+        message: ErrorMessages.ITEM_NOT_FOUND,
+      });
+    }
+
+    return createResponse(StatusCodes.OK, item);
+  } catch (error) {
+    return createResponse(StatusCodes.NOT_FOUND, {
+      message: ErrorMessages.INTERNAL_SERVER_ERROR,
+    });
+  }
 };
 
 const getAllItems = async () => {
+  console.log("getAllItems controller");
   const items = await itemService.getAll();
-
-  return createResponse(StatusCodes.OK, items);
+  console.log("items", items);
+  try {
+    return createResponse(StatusCodes.OK, items);
+  } catch (error) {
+    return createResponse(StatusCodes.INTERNAL_SERVER_ERROR, {
+      message: ErrorMessages.INTERNAL_SERVER_ERROR,
+    });
+  }
 };
 
 const updateItem = async (event: APIGatewayEvent) => {
@@ -41,10 +66,24 @@ const updateItem = async (event: APIGatewayEvent) => {
     });
   }
 
-  const itemData = JSON.parse(event.body ?? "{}");
-  const updatedItem = await itemService.update(itemId, itemData);
+  const existingItem = await itemService.get(itemId);
 
-  return createResponse(StatusCodes.OK, updatedItem);
+  if (!existingItem) {
+    return createResponse(StatusCodes.NOT_FOUND, {
+      message: ErrorMessages.ITEM_NOT_FOUND,
+    });
+  }
+
+  const itemData = JSON.parse(event.body ?? "{}");
+
+  try {
+    const updatedItem = await itemService.update(itemId, itemData);
+    return createResponse(StatusCodes.OK, updatedItem);
+  } catch (error) {
+    return createResponse(StatusCodes.BAD_REQUEST, {
+      message: ErrorMessages.INVALID_DATA,
+    });
+  }
 };
 
 const deleteItem = async (event: APIGatewayEvent) => {
@@ -56,10 +95,22 @@ const deleteItem = async (event: APIGatewayEvent) => {
     });
   }
 
-  const itemDeleted = await itemService.get(itemId);
-  await itemService.delete(itemId);
+  const existingItem = await itemService.get(itemId);
 
-  return createResponse(StatusCodes.NO_CONTENT, itemDeleted);
+  if (!existingItem) {
+    return createResponse(StatusCodes.NOT_FOUND, {
+      message: ErrorMessages.ITEM_NOT_FOUND,
+    });
+  }
+
+  try {
+    await itemService.delete(itemId);
+    return createResponse(StatusCodes.OK, existingItem);
+  } catch (error) {
+    return createResponse(StatusCodes.INTERNAL_SERVER_ERROR, {
+      message: ErrorMessages.INTERNAL_SERVER_ERROR,
+    });
+  }
 };
 
 export { createItem, getItem, getAllItems, updateItem, deleteItem };
